refactor(exec): reuse exec handler in ffmpeg handler
