+++ pkg/define/const.go
@@ -3,6 +3,7 @@ package define
 const (
 	FFMPEG        = "ffmpeg"
 	FFPROBE       = "ffprobe"
+	SHELL         = "bash"
 	InstallFFMPEG = "install_ffmpeg"
 	StudioDir     = ".oomol-studio"
 	FFReleaseURL  = "https://github.com/oomol/builded/releases/download/v1.4/ffmpeg_macos_arm64_ventura.tar.xz"
