+++ pkg/ffmpeg/runner.go
@@ -3,7 +3,7 @@ package ffmpeg
 import (
 	"context"
 	"github.com/gliderlabs/ssh"
-	"os/exec"
+	"sshd/pkg/handler"
 )
 
 type Runner struct {
@@ -14,17 +14,5 @@ type Runner struct {
 }
 
 func (r *Runner) Run(ctx context.Context) error {
-	cmd := exec.CommandContext(ctx, r.File, r.Args...)
-	cmd.Env = append(cmd.Env, r.Envs...)
-	cmd.Stdout = r.Session
-	cmd.Stderr = r.Session.Stderr()
-	if err := cmd.Start(); err != nil {
-		return err
-	}
-
-	if err := cmd.Wait(); err != nil {
-		return err
-	}
-
-	return nil
+	return handler.ExecHandler(r.Session, ctx, r.File, r.Envs, r.Args)
 }
